DEV-37860; refactor: move ai service to a separate module
