+++ src/services/ai-service/ai-service.types.ts
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2024-present Acrolinx GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { CommonIssue } from '../../check';
+
+export interface AiFeatures {
+  ai: boolean;
+  aiAssistant: boolean;
+}
+
+export interface IsAIEnabledInformation {
+  tenant: string;
+  value: boolean;
+  userHasPrivilege: boolean;
+}
+
+export type ChatCompletionRequest = {
+  issue: CommonIssue;
+  count: number;
+  targetUuid: string;
+};
+
+export interface WriteResponse {
+  response: string;
+}
+
+export interface AIServiceError {
+  code: number;
+  message: string;
+}
+
+export enum AIServiceErrorTypes {
+  AI_SERVICE_ERROR = 'AI_SERVICE_ERROR',
+}
